+++ activerecord/test/models/pirate.rb
@@ -27,6 +27,8 @@ class Pirate < ActiveRecord::Base
   has_many :treasure_estimates, through: :treasures, source: :price_estimates
 
   has_one :ship
+  has_one :ship_with_autosave, class_name: "Ship", autosave: true
+  has_one :ship_without_autosave, class_name: "Ship", autosave: false
   has_one :update_only_ship, class_name: "Ship"
   has_one :non_validated_ship, class_name: "Ship"
   has_many :birds, -> { order("birds.id ASC") }